Converts site to typescript
Converts files, adds interfaces, ensures pages are working
