+++ src/components/Grid/Col/Col.tsx
@@ -1,27 +1,24 @@
 /** @jsx jsx */
-// import React from "react"
 import { jsx } from 'theme-ui'
 
-const Col = ({ xs, sm, md, lg, xl, className, children }, props) => {
-  const calcWidth = colWidth => {
-    if (colWidth === undefined) {
-      return null
-    }
+type colSizeType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12
 
-    return `${(colWidth / 12) * 100}%`
-  }
+interface propTypes {
+  xs?: colSizeType
+  sm?: colSizeType
+  md?: colSizeType
+  lg?: colSizeType
+  xl?: colSizeType
+  sx?: {}
+  className?: string
+  children: any
+}
 
+const Col = ({ xs, sm, md, lg, xl, sx, className, children }: propTypes) => {
   /**
    * We want the 'flex' css property to return '1 0 xx%' unless a colWidth has not been set
    * in which case we only want the first result to return '1 0 100%' (to ensure all boxes are fully stretched)
    */
-  const calcFlex = colWidth => {
-    if (colWidth === undefined) {
-      return null
-    }
-
-    return `1 0 ${(colWidth / 12) * 100}%`
-  }
 
   return (
     <div
@@ -52,3 +49,19 @@ const Col = ({ xs, sm, md, lg, xl, className, children }, props) => {
 }
 
 export default Col
+
+const calcWidth = (colWidth: colSizeType | undefined) => {
+  if (colWidth === undefined) {
+    return null
+  }
+
+  return `${(colWidth / 12) * 100}%`
+}
+
+const calcFlex = (colWidth: colSizeType | undefined) => {
+  if (colWidth === undefined) {
+    return null
+  }
+
+  return `1 0 ${(colWidth / 12) * 100}%`
+}
