+++ src/components/Grid/index.ts
@@ -1,0 +1,4 @@
+export { default as Container } from './Container'
+export { default as Row } from './Row'
+export { default as Col } from './Col'
+// export { default as Grid } from './Grid'
