+++ src/components/Layout/Global/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Global'
