+++ src/components/Layout/Header/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Header'
+export { default as DarkButton } from './DarkButton'
