+++ src/components/Layout/index.ts
@@ -1,0 +1,4 @@
+export { default } from './Layout'
+export { default as Footer } from './Footer'
+export { default as Global } from './Global'
+export { default as Header } from './Header'
