+++ src/components/Post/index.ts
