+++ src/components/Seo/Seo.tsx
@@ -1,15 +1,14 @@
-/**
- * SEO component that queries for data with
- *  Gatsby's useStaticQuery React hook
- *
- * See: https://www.gatsbyjs.org/docs/use-static-query/
- */
-
 import React from 'react'
 import Helmet from 'react-helmet'
 import { useStaticQuery, graphql } from 'gatsby'
 
-const SEO = ({ title, description, meta }) => {
+interface propTypes {
+  title: string
+  description?: string
+  meta?: []
+}
+
+const SEO = ({ title, description, meta }: propTypes) => {
   const { site } = useStaticQuery(
     graphql`
       query {
