--- src/components/index.js
@@ -1,10 +1,0 @@
-export { default as Col } from './Col'
-export { default as Container } from './Container'
-export { default as DarkButton } from './DarkButton'
-export { default as Header } from './Header'
-export { default as Hero } from './Hero'
-export { default as Layout } from './Layout'
-export { default as Post } from './Post'
-export { default as Row } from './Row'
-export { default as SEO } from './Seo'
-export { default as Socials } from './Socials'
