+++ src/components/index.ts
@@ -1,0 +1,6 @@
+// export { default as Grid } from './Grid'
+export { default as Hero } from './Hero'
+export { default as Layout } from './Layout'
+export { default as Post } from './Post'
+export { default as SEO } from './Seo'
+export { default as Socials } from './Socials'
